sui-adapter: move unit tests into dedicated module
