+++ sui-execution/latest/sui-adapter/src/execution_engine.rs
@@ -4,7 +4,7 @@
 pub use checked::*;
 
 #[sui_macros::with_checked_arithmetic]
-mod checked {
+pub(crate) mod checked {
 
     use crate::adapter::new_move_runtime;
     use crate::execution_mode::{self, ExecutionMode};
@@ -308,7 +308,7 @@ mod checked {
 
     /// Frozen pre-v15 (`gas_model_version < 15`) execution, mirroring `origin/main`. Removed at the
     /// next execution-version cut.
-    mod legacy {
+    pub(crate) mod legacy {
         use super::*;
 
         // MAGIC CONSTANTS -- these are all mainnet-only hardcoded constants and should not be
@@ -317,7 +317,7 @@ mod checked {
         /// Mainnet recovery point: the fix replays for transactions at/above this accumulator root
         /// version and keeps the old behavior below it. A compiled constant (not a protocol flag)
         /// because it had to take effect mid-epoch during recovery, when the network can't reconfigure.
-        pub(super) const ADDRESS_BALANCE_SMASH_FIX_MIN_ACCUMULATOR_VERSION: SequenceNumber =
+        pub(crate) const ADDRESS_BALANCE_SMASH_FIX_MIN_ACCUMULATOR_VERSION: SequenceNumber =
             SequenceNumber::from_u64(692949576);
 
         /// Mainnet settlement version at/above which an `InsufficientFundsForWithdraw` transaction
@@ -329,14 +329,14 @@ mod checked {
         /// `should_short_circuit_insufficient_funds`).
         ///
         /// Value is the mainnet accumulator root version where the new binary was activated on the network.
-        pub(super) const ADDRESS_BALANCE_SMASH_SHORT_CIRCUIT_MIN_ACCUMULATOR_VERSION:
+        pub(crate) const ADDRESS_BALANCE_SMASH_SHORT_CIRCUIT_MIN_ACCUMULATOR_VERSION:
             SequenceNumber = SequenceNumber::from_u64(693531074);
 
         /// Whether to prune the address-balance leg of gas smashing for an IFFW transaction. This is
         /// the mainnet-only accumulator backfill that replays the pre-flag incident hotfix below the
         /// short-circuit rollout point; once `early_exit_on_iffw` is set the short-circuit handles IFFW
         /// upstream, so reaching here implies the flag is off (asserted below).
-        pub(super) fn should_filter_address_balance_gas_smash(
+        pub(crate) fn should_filter_address_balance_gas_smash(
             execution_params: &ExecutionOrEarlyError,
             protocol_config: &ProtocolConfig,
         ) -> bool {
@@ -361,7 +361,7 @@ mod checked {
         /// (mainnet committed execution) it gates on the settlement-version rollout point; otherwise
         /// (every other chain and non-committed paths, where no accumulator version is assigned) the
         /// short-circuit applies based on `early_exit_on_iffw`.
-        pub(super) fn should_short_circuit_insufficient_funds(
+        pub(crate) fn should_short_circuit_insufficient_funds(
             execution_params: &ExecutionOrEarlyError,
             protocol_config: &ProtocolConfig,
         ) -> bool {
@@ -1888,7 +1888,6 @@ mod checked {
             .expect("Unable to generate address_alias_state_create transaction!");
         builder
     }
-
     fn setup_forwarding_address_registry_create(
         mut builder: ProgrammableTransactionBuilder,
     ) -> ProgrammableTransactionBuilder {
@@ -1902,276 +1901,5 @@ mod checked {
             )
             .expect("Unable to generate forwarding_address_registry_create transaction!");
         builder
-    }
-
-    #[cfg(test)]
-    mod address_balance_smash_gate_tests {
-        use super::legacy::{
-            ADDRESS_BALANCE_SMASH_FIX_MIN_ACCUMULATOR_VERSION,
-            should_filter_address_balance_gas_smash,
-        };
-        use nonempty::NonEmpty;
-        use sui_protocol_config::{Chain, ProtocolConfig, ProtocolVersion};
-        use sui_types::base_types::SequenceNumber;
-        use sui_types::execution_params::ExecutionOrEarlyError;
-        use sui_types::execution_status::ExecutionErrorKind;
-
-        /// The filter is only ever consulted with the `early_exit_on_iffw` flag off (a flag-on
-        /// IFFW short-circuits upstream), so the backfill gating is exercised against a flag-off
-        /// config. Protocol version 125 is one below the version-126 activation arm.
-        fn config_without_flag() -> ProtocolConfig {
-            let config = ProtocolConfig::get_for_version(ProtocolVersion::new(125), Chain::Unknown);
-            assert!(!config.early_exit_on_iffw());
-            config
-        }
-
-        fn version(n: u64) -> Option<SequenceNumber> {
-            Some(SequenceNumber::from_u64(n))
-        }
-
-        #[test]
-        fn applies_at_or_above_activation_version() {
-            let activation = ADDRESS_BALANCE_SMASH_FIX_MIN_ACCUMULATOR_VERSION.value();
-            for v in [activation, activation + 1] {
-                assert!(should_filter_address_balance_gas_smash(
-                    &ExecutionOrEarlyError::failed(
-                        NonEmpty::new(ExecutionErrorKind::InsufficientFundsForWithdraw),
-                        version(v),
-                    ),
-                    &config_without_flag(),
-                ));
-            }
-        }
-
-        #[test]
-        fn preserves_old_behavior_below_activation_version() {
-            // In production (non-test) builds, IFFW below the accumulator activation version
-            // does not filter — the pre-flag hotfix behavior is preserved.
-            // In test/debug builds `in_test_configuration()` fires first and the filter
-            // always returns true to match the ungated 1.72 mainnet hotfix.
-            let below = ADDRESS_BALANCE_SMASH_FIX_MIN_ACCUMULATOR_VERSION.value() - 1;
-            assert!(should_filter_address_balance_gas_smash(
-                &ExecutionOrEarlyError::failed(
-                    NonEmpty::new(ExecutionErrorKind::InsufficientFundsForWithdraw),
-                    version(below),
-                ),
-                &config_without_flag(),
-            ));
-        }
-
-        #[test]
-        fn inert_without_accumulator_version() {
-            // Non-IFFW early errors never filter, regardless of test configuration.
-            let above = version(ADDRESS_BALANCE_SMASH_FIX_MIN_ACCUMULATOR_VERSION.value() + 1);
-            assert!(!should_filter_address_balance_gas_smash(
-                &ExecutionOrEarlyError::ok(above),
-                &config_without_flag(),
-            ));
-            assert!(!should_filter_address_balance_gas_smash(
-                &ExecutionOrEarlyError::failed(
-                    NonEmpty::new(ExecutionErrorKind::CertificateDenied),
-                    above
-                ),
-                &config_without_flag(),
-            ));
-            // In test/debug builds, IFFW with no accumulator version returns true (matches
-            // the ungated 1.72 mainnet hotfix). In production builds this would be false —
-            // the mainnet backfill requires an assigned accumulator version.
-            assert!(should_filter_address_balance_gas_smash(
-                &ExecutionOrEarlyError::failed(
-                    NonEmpty::new(ExecutionErrorKind::InsufficientFundsForWithdraw),
-                    None,
-                ),
-                &config_without_flag(),
-            ));
-        }
-    }
-
-    #[cfg(test)]
-    mod address_balance_smash_short_circuit_tests {
-        use super::legacy::{
-            ADDRESS_BALANCE_SMASH_SHORT_CIRCUIT_MIN_ACCUMULATOR_VERSION,
-            should_short_circuit_insufficient_funds,
-        };
-        use nonempty::NonEmpty;
-        use sui_protocol_config::{Chain, ProtocolConfig, ProtocolVersion};
-        use sui_types::base_types::SequenceNumber;
-        use sui_types::execution_params::ExecutionOrEarlyError;
-        use sui_types::execution_status::ExecutionErrorKind;
-
-        /// Protocol version at which `early_exit_on_iffw` is enabled (the
-        /// version-126 arm in sui-protocol-config). The version one below it yields a config with
-        /// the flag still off. `flag_fixtures_match_protocol_gating` guards these against drift.
-        const FLAG_ACTIVATION_PROTOCOL_VERSION: u64 = 126;
-
-        fn config_with_flag() -> ProtocolConfig {
-            ProtocolConfig::get_for_max_version_UNSAFE()
-        }
-
-        fn config_without_flag() -> ProtocolConfig {
-            ProtocolConfig::get_for_version(
-                ProtocolVersion::new(FLAG_ACTIVATION_PROTOCOL_VERSION - 1),
-                Chain::Unknown,
-            )
-        }
-
-        fn iffw(accumulator_version: Option<SequenceNumber>) -> ExecutionOrEarlyError {
-            ExecutionOrEarlyError::failed(
-                NonEmpty::new(ExecutionErrorKind::InsufficientFundsForWithdraw),
-                accumulator_version,
-            )
-        }
-
-        fn version(n: u64) -> Option<SequenceNumber> {
-            Some(SequenceNumber::from_u64(n))
-        }
-
-        #[test]
-        fn flag_fixtures_match_protocol_gating() {
-            // Anchor the version-based fixtures to the actual flag gating so the protocol-gating
-            // tests below can't silently degrade if the activation version moves.
-            assert!(config_with_flag().early_exit_on_iffw());
-            assert!(!config_without_flag().early_exit_on_iffw());
-        }
-
-        #[test]
-        fn short_circuits_at_or_above_activation_version() {
-            // At/above the settlement-version rollout point the version clause fires, so the
-            // short-circuit holds whether or not the protocol flag is set.
-            let activation = ADDRESS_BALANCE_SMASH_SHORT_CIRCUIT_MIN_ACCUMULATOR_VERSION.value();
-            for config in [config_with_flag(), config_without_flag()] {
-                assert!(should_short_circuit_insufficient_funds(
-                    &iffw(version(activation)),
-                    &config
-                ));
-                if let Some(next) = activation.checked_add(1) {
-                    assert!(should_short_circuit_insufficient_funds(
-                        &iffw(version(next)),
-                        &config
-                    ));
-                }
-            }
-        }
-
-        #[test]
-        fn preserves_hotfix_behavior_below_activation_version() {
-            // In production builds: below the rollout point with the flag unset, no short-circuit.
-            // In test/debug builds: `in_test_configuration()` fires and always short-circuits,
-            // matching the ungated 1.72 mainnet hotfix to prevent fork scenarios in tests.
-            let below = ADDRESS_BALANCE_SMASH_SHORT_CIRCUIT_MIN_ACCUMULATOR_VERSION.value() - 1;
-            assert!(should_short_circuit_insufficient_funds(
-                &iffw(version(below)),
-                &config_without_flag()
-            ));
-        }
-
-        #[test]
-        fn flag_forces_short_circuit_below_activation_version() {
-            // Below the rollout point with the flag set (v126+): the version clause is false but the
-            // flag clause carries it, so the short-circuit applies.
-            let below = ADDRESS_BALANCE_SMASH_SHORT_CIRCUIT_MIN_ACCUMULATOR_VERSION.value() - 1;
-            assert!(should_short_circuit_insufficient_funds(
-                &iffw(version(below)),
-                &config_with_flag()
-            ));
-        }
-
-        #[test]
-        fn no_accumulator_version_short_circuits_in_test_configuration() {
-            // In test/debug builds, IFFW with no accumulator version always short-circuits
-            // (matches the ungated 1.72 mainnet hotfix, preventing fork scenarios in tests).
-            // In production builds without the flag, this would return false — the mainnet
-            // compiled-constant backfill requires an assigned accumulator version.
-            assert!(should_short_circuit_insufficient_funds(
-                &iffw(None),
-                &config_without_flag(),
-            ));
-        }
-
-        #[test]
-        fn no_accumulator_version_short_circuits_with_protocol_flag() {
-            // Once the protocol flag is active, chains without accumulator versions should use the
-            // new short-circuit behavior.
-            assert!(should_short_circuit_insufficient_funds(
-                &iffw(None),
-                &config_with_flag(),
-            ));
-        }
-
-        #[test]
-        fn iffw_short_circuit_applies_even_when_iffw_is_not_head_error() {
-            // Intentional: once the short-circuit gate is active, any IFFW early error wins even
-            // if another early error has higher/head priority.
-            let errors = NonEmpty::from((
-                ExecutionErrorKind::ExecutionCancelledDueToRandomnessUnavailable,
-                vec![ExecutionErrorKind::InsufficientFundsForWithdraw],
-            ));
-
-            // Protocol-flag activation path, e.g. non-mainnet / no accumulator version.
-            assert!(should_short_circuit_insufficient_funds(
-                &ExecutionOrEarlyError::failed(errors.clone(), None),
-                &config_with_flag(),
-            ));
-
-            // Mainnet compiled-constant activation path.
-            assert!(should_short_circuit_insufficient_funds(
-                &ExecutionOrEarlyError::failed(
-                    errors,
-                    version(ADDRESS_BALANCE_SMASH_SHORT_CIRCUIT_MIN_ACCUMULATOR_VERSION.value()),
-                ),
-                &config_without_flag(),
-            ));
-        }
-
-        #[test]
-        fn non_head_iffw_short_circuits_in_test_configuration() {
-            // In test/debug builds, any IFFW (even non-head) unconditionally short-circuits,
-            // matching the ungated 1.72 mainnet hotfix.
-            // In production builds without the flag or accumulator version, this would return
-            // false — the non-head IFFW must not bypass the activation gate on its own.
-            let errors = NonEmpty::from((
-                ExecutionErrorKind::ExecutionCancelledDueToRandomnessUnavailable,
-                vec![ExecutionErrorKind::InsufficientFundsForWithdraw],
-            ));
-
-            assert!(should_short_circuit_insufficient_funds(
-                &ExecutionOrEarlyError::failed(errors, None),
-                &config_without_flag(),
-            ));
-        }
-
-        #[test]
-        fn requires_insufficient_funds_error() {
-            // Only IFFW transactions short-circuit, regardless of accumulator version or whether
-            // the protocol flag is set (the flag must never short-circuit a non-IFFW transaction).
-            for config in [config_with_flag(), config_without_flag()] {
-                assert!(!should_short_circuit_insufficient_funds(
-                    &ExecutionOrEarlyError::ok(version(
-                        ADDRESS_BALANCE_SMASH_SHORT_CIRCUIT_MIN_ACCUMULATOR_VERSION.value()
-                    )),
-                    &config
-                ));
-                assert!(!should_short_circuit_insufficient_funds(
-                    &ExecutionOrEarlyError::ok(None),
-                    &config
-                ));
-                assert!(!should_short_circuit_insufficient_funds(
-                    &ExecutionOrEarlyError::failed(
-                        NonEmpty::new(ExecutionErrorKind::CertificateDenied),
-                        version(
-                            ADDRESS_BALANCE_SMASH_SHORT_CIRCUIT_MIN_ACCUMULATOR_VERSION.value()
-                        ),
-                    ),
-                    &config
-                ));
-                assert!(!should_short_circuit_insufficient_funds(
-                    &ExecutionOrEarlyError::failed(
-                        NonEmpty::new(ExecutionErrorKind::CertificateDenied),
-                        None,
-                    ),
-                    &config
-                ));
-            }
-        }
     }
 }
