+++ sui-execution/latest/sui-adapter/src/static_programmable_transactions/loading/ast.rs
@@ -401,12 +401,3 @@ impl TryFrom<&Datatype> for StructTag {
         })
     }
 }
-
-//**************************************************************************************************
-// Tests
-//**************************************************************************************************
-
-#[test]
-fn enum_size() {
-    assert_eq!(std::mem::size_of::<Type>(), 16);
-}
