+++ sui-execution/latest/sui-adapter/src/unit_tests/mod.rs
@@ -1,0 +1,6 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+mod address_balance_smash_gate_tests;
+mod address_balance_smash_short_circuit_tests;
+mod ast_tests;
